+++ src/routers/login.js
@@ -1,6 +1,7 @@
-const express = require('express')
-const User = require('../models/user')
-const router = new express.Router()
+const express = require('express');
+const User = require('../models/user');
+const auth = require('../middleware/auth');
+const router = new express.Router();
 
 // Rotas
 router.get("/", function(req, res) {
@@ -15,47 +16,84 @@ router.get("/cadastro", function(req, res) {
   res.render('cadastro');
 });
 
-router.post("/logar", async (req, res) => {
-  let username = req.body.username;
-  let password = req.body.password;
+// router.post("/logar", async (req, res) => {
+//   let username = req.body.username;
+//   let password = req.body.password;
 
-  if(!username || !password) {
-    return res.render('login', {
+//   if(!username || !password) {
+//     return res.render('login', {
+//       message: true,
+//       alertMessage: 'Nome de usuário ou senha incorretos'
+//     });
+//   }
+
+//   try {
+//     const users = await User.find({ username: username, password: password });
+
+//     if(users.length === 0) {
+//       return res.render('login', {
+//         message: true,
+//         alertMessage: 'Nome de usuário ou senha incorretos'
+//       });
+//     }
+//     return res.render('pagina-inicial', {
+//       name: username,
+//     });
+//     // res.send(users);
+
+//   } catch (e) {
+//     res.status(500).send();
+//   }
+
+//   // if(username==='admin' && password==='admin') {
+//   //   res.render('pagina-inicial', {
+//   //     name: username
+//   //   });
+//   // } else {
+//   //   res.render('login');
+//   // }
+// });
+
+router.post('/logar', async (req, res) => {
+  try {
+    const user = await User.findByCredentials(req.body.username, req.body.password);
+    const token = await user.generateAuthToken();
+    res.send({ user, token });
+  } catch (e) {
+    //res.status(400).send('Nome de usuário ou senha incorretos');
+    res.render('login', {
       message: true,
       alertMessage: 'Nome de usuário ou senha incorretos'
     });
   }
+})
 
-  try {
-    const users = await User.find({ username: username, password: password });
+// router.post("/cadastrar", async (req, res) => {
+//   if (req.body['password-check'] !== req.body.password) {
+//     // return res.status(400).send('As senhas diferem!')
+//     return res.render('cadastro', {
+//       message: true,
+//       alertMessage: 'As senhas diferem!'
+//     });
+//   }
+//   const user = new User(req.body);
 
-    if(users.length === 0) {
-      return res.render('login', {
-        message: true,
-        alertMessage: 'Nome de usuário ou senha incorretos'
-      });
-    }
-    return res.render('pagina-inicial', {
-      name: username,
-    });
-    // res.send(users);
+//   try {
+//     await user.save();
+//     res.status(201).send(user);
+//   } catch (e) {
+//     // res.status(400).send(e);
+//     return res.render('cadastro', {
+//       message: true,
+//       alertMessage: 'Nome de usuário ou senha inválidos'
+//     });
+//   }
 
-  } catch (e) {
-    res.status(500).send();
-  }
+//   // res.render('login');
+// });
 
-  // if(username==='admin' && password==='admin') {
-  //   res.render('pagina-inicial', {
-  //     name: username
-  //   });
-  // } else {
-  //   res.render('login');
-  // }
-});
-
-router.post("/cadastrar", async (req, res) => {
+router.post('/cadastrar', async (req, res) => {
   if (req.body['password-check'] !== req.body.password) {
-    // return res.status(400).send('As senhas diferem!')
     return res.render('cadastro', {
       message: true,
       alertMessage: 'As senhas diferem!'
@@ -65,7 +103,8 @@ router.post("/cadastrar", async (req, res) => {
 
   try {
     await user.save();
-    res.status(201).send(user);
+    const token = await user.generateAuthToken();
+    res.status(201).send({ user, token });
   } catch (e) {
     // res.status(400).send(e);
     return res.render('cadastro', {
@@ -73,8 +112,26 @@ router.post("/cadastrar", async (req, res) => {
       alertMessage: 'Nome de usuário ou senha inválidos'
     });
   }
+});
 
-  // res.render('login');
+router.get('/me', auth, async (req, res) => {
+  res.send(req.user);
+});
+
+router.post('/sair', auth, async (req, res) => {
+  try {
+    req.user.tokens = req.user.tokens.filter((token) => {
+      return token.token !== req.token;
+    })
+    await req.user.save();
+
+    res.render('login', {
+      message: true,
+      alertMessage: 'Desconectado com sucesso!'
+    });
+  } catch (e) {
+    res.status(500).send();
+  }
 });
 
 module.exports = router;